+++ WordleWinUI/App.xaml.cpp
@@ -16,6 +16,7 @@ namespace winrt
 
 winrt::Microsoft::UI::Xaml::Window winrt::App::window{ nullptr };
 
+
 /// <summary>
 /// Initializes the singleton application object.  This is the first line of authored code
 /// executed, and as such is the logical equivalent of main() or WinMain().